Seperate the different functions in previous "tree_detector.cpp" into 1) laser part, 2) vision part, 3) target tree tracking part
