+++ jsk_uav_forest_perception/CMakeLists.txt
@@ -19,7 +19,7 @@ find_package(OpenCV REQUIRED)
 message(WARNING "OPENCV ${OpenCV_VERSION} FOUND")
 
 generate_dynamic_reconfigure_options(
-  cfg/ColorRegionDetector.cfg
+  cfg/HsvFilter.cfg
   )
 
 
@@ -40,18 +40,18 @@ include_directories(
   ${catkin_INCLUDE_DIRS}
 )
 
+add_library(laser_clustering_filter SHARED src/laser_clustering_filter.cpp)
+target_link_libraries(laser_clustering_filter ${catkin_LIBRARIES})
 
-add_library(color_region_detection src/color_region_detector.cpp)
-target_link_libraries(color_region_detection ${catkin_LIBRARIES})
-add_dependencies(color_region_detection ${PROJECT_NAME}_gencfg)
+add_library(vision_detection src/vision_detector.cpp)
+target_link_libraries(vision_detection ${catkin_LIBRARIES})
+add_dependencies(vision_detection ${PROJECT_NAME}_gencfg)
 
-add_library(jsk_laser_scan_filters SHARED src/laser_scan_filters.cpp)
-target_link_libraries(jsk_laser_scan_filters ${catkin_LIBRARIES})
 
-add_library(tree_detector src/tree_detector.cpp)
-target_link_libraries(tree_detector ${catkin_LIBRARIES})
-add_executable(tree_detector_node src/tree_detector_node.cpp)
-target_link_libraries (tree_detector_node ${catkin_LIBRARIES} tree_detector)
+add_library(tree_tracking src/tree_tracking.cpp)
+target_link_libraries(tree_tracking ${catkin_LIBRARIES})
+add_executable(tree_tracking_node src/tree_tracking_node.cpp)
+target_link_libraries (tree_tracking_node ${catkin_LIBRARIES} tree_tracking)
 
 
 #############
