+++ jsk_uav_forest_perception/cfg/HsvFilter.cfg
@@ -13,4 +13,4 @@ gen.add("color_hsv_s_max",    int_t,    0, "S upper bound", 255,  0, 255)
 gen.add("color_hsv_v_min",    int_t,    0, "V lower bound", 2,  0, 255)
 gen.add("color_hsv_v_max",    int_t,    0, "V upper bound", 216,  0, 255)
 
-exit(gen.generate(PACKAGE, "jsk_uav_forest_perception", "ColorRegionDetector"))
+exit(gen.generate(PACKAGE, "jsk_uav_forest_perception", "HsvFilter"))
