+++ jsk_uav_forest_perception/laser_filters_plugins.xml
@@ -1,9 +1,9 @@
 <class_libraries>
-  <library path="lib/libjsk_laser_scan_filters">
-    <class name="laser_filters/ScanTreeFilter" type="laser_filters::ScanTreeFilter"
+  <library path="lib/liblaser_clustering_filter">
+    <class name="laser_filters/LaserClusteringFilter" type="laser_filters::LaserClusteringFilter"
 	    base_class_type="filters::FilterBase<sensor_msgs::LaserScan>">
       <description>
-	This is a filter which extract tree from a laser.
+	This is a filter which do clustering from a laser.
       </description>
     </class>
   </library>
