+++ jsk_uav_forest_perception/launch/laser_process.launch
@@ -7,12 +7,11 @@
     <rosparam command="load" file="$(find jsk_uav_forest_perception)/param/LaserShadowFilter.yaml" />
   </node>
 
-  <!-- foot object laser filter -->
-<!--
+  <!-- laser clustering  filter -->
   <node pkg="laser_filters" type="scan_to_scan_filter_chain" output="screen" name="trees_filter">
     <remap from="scan" to="shadow_filtered" />
-    <remap from="scan_filtered" to="scan_filtered_tree" />
+    <remap from="scan_filtered" to="scan_clustered" />
     <rosparam command="load" file="$(find jsk_uav_forest_perception)/param/TreeFilter.yaml" />
-    </node>
--->
+
+  </node>
 </launch>
