+++ jsk_uav_forest_perception/param/LaserShadowFilter.yaml
@@ -2,7 +2,7 @@ scan_filter_chain:
 - name: range
   type: laser_filters/LaserScanRangeFilter
   params:
-        lower_threshold: 0
+        lower_threshold: 0.3
         upper_threshold: 15
 - name: shadows
   type: laser_filters/ScanShadowsFilter
